Added support for ask.fm
Fixes #151. As ask.fm constantly keeps deleting and recreating it's
body, #hzImg is constantly being removed (and not added back in, as the
internal data structure is not updated accordingly). Because the whole
body is being removed, I added a MutationObserver that observes the
html-element for changes, and if #hzImg has been removed the data
structure is updated acordingly (hoverzoom.hzImg is set to false).

I also removed two tabstops in manifest.json, and replaced them with
spaces.