Delay Secondary Request
FetLife's feed can have LOTS of thumbnails next to each other (someone can like 50+ photos, and it'll show them in a 7xWhatever grid).  Previously, it was really easy to sweep a bunch of images with the mouse to trigger lots of requests, both hammering and FetLife and causing HoverZoom to flicker lots of different full size images.

Not sure how to address the larger flicker/UI issue when that was happening, so just delayed the prepareFromDocument() call.