Enhancements to Pinterest plug-in & fix for HZ+ function prepareDownscaledImages()
Enhancements to Pinterest plug-in:
- enable zoom on avatars & background images
- fix for images without srcset

Fix for HZ+ function prepareDownscaledImages()
Plug-ins and this function add srcs to the same array: img.data().hoverZoomSrc
Since downscaled images srcs are of lower quality than srcs provided by plug-ins, they should be added to the bottom of array (= pushed, instead of unshifted).
So they'll be displayed only if srcs from plug-ins are all invalid.