New warning message when page is not focused
Sometimes action keys seem to be broken, this usually happens when page has lost focus.
A warning message will ask user to click anywhere on page so it has focus and action keys can work.
This message is displayed only once (until page is reloaded), so users that do not use action keys are not bothered.