Add a link to the list of plugins in the readme
The list of supported sites is not easy to find. This is obvious to you as the author, but I needed to go from the addons page to the repository, read the readme, enter a query into the search, and only on the second page of results I understand that each supported site is a separate file in the plugins folder. It took some time. And not every addon user will be able to find it. But I think a lot of people will.
This PR is just a simple example of how you can help the user find the list. I would recommend that you provide a link to such information on the addon page. You can also add a list opening from an already installed addon.