New action key: ban an image, video or audio track
A new action key (default value: "B") can be used to prevent an image, video or audio track from being zoomed (and played) next time it is hovered.
Ban is not lifted by just restarting browser. So in Advanced options, a new button is available in order to reset banned items so they can be zoomed again.