Fix right click not opening context menu when set to lock image  (#1529)
This fixes context menu not showing up if right click is set to "Lock
Image" and the image is currently locked. It should function like it use
to now.

- fixes #1487

Also, I'm not sure where to report this, but the Storage API migration
broke saving and the options UI on Firefox. I'm working on a fix for it,
but I am not able to figure out how to get storage.sync.set to work on
firefox yet