Change toolbar icon when on zoomable sites (#1583)
The deprecated code once used to put the icon in the address bar can now
be used to change the icon of the toolbar icon instead. This PR reworks
the code to add that functionality. Idea came from
https://github.com/extesy/hoverzoom/issues/1578#issuecomment-2744565717:
- on sites with zoomable images, icon turns a purplish blue. Is similar
to how the icon use to show in the address bar when on zoomable web
pages
- updated option's message text to reflect new function

I wasn't sure what to make the new icon, so if you have a better idea
for it, I'd be happy to change it