Fix auto lock (#1594)
This fixes auto lock so it behaves as if one pressed the "Lock Image"
key.

With auto lock on now, the image:
- fits to screen, unless "Locked images start zoomed in" is enabled
- will toggle between zoom states with additional presses of the "Lock
Image" key
- uses the zoom factor for zoom amount unless "Use zoom factor for lock
image" is disabled

I also fixed "Locked images start zoomed in" not working in general. 

Likely resolves #1586 (Not entirely sure the issue in that, but that's
where I discovered the bugs fixed in this PR)

---------

Co-authored-by: Oleg Anashkin <github@oleg.anashkin.org>