Add permissions checking function to background (#1600)
This allows hoverzoom.js to check for 'downloads' permission before
saving an image, in case the permission was already active or changed
outside the options page. This also allows an alert to be sent from the
main script. The permission could instead be checked in hoverzoom.js
```init()``` so it's not called every time a user tries to save an image