Fix copy image and copy image url (#1608)
This attempts to fix "Copy Image" and "Copy Image URL" so it works in
Chrome and Firefox. The main issue is Chrome requires
"Access-Control-Allow-Origin" headers to be set. This is infeasible as
it would have to be set for every site, unless it is set when the "Copy
Image" key is pressed
- will fix #1607 once cors issue is resolved