Added an option to disable looping albums (#1620)
The new option, `galleriesLoopEnabled`, is true by default. If disabled,
navigating past the beginning or end of an album will do nothing, rather
than looping back to the other side of it.

I created this PR after I realized that I could just do
https://github.com/extesy/hoverzoom/issues/1618 myself. :)

Unfortunately, I have no idea how to create all the translations for the
option description, so that'll have to be done before this PR can be
accepted. Is there some script I can run that will auto-translate the
`en` locale message that I did write?